Address runtime type export review feedback
